вынес конструктор размера пиццы в отдельный компонент
